Migrating ShapeProfiles to Geometry_oM
