+++ Geometry_oM/Geometry_oM.csproj
@@ -62,6 +62,23 @@
     <Compile Include="Misc\Tolerance.cs" />
     <Compile Include="Misc\TransformMatrix.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ShapeProfiles\AngleProfile.cs" />
+    <Compile Include="ShapeProfiles\BoxProfile.cs" />
+    <Compile Include="ShapeProfiles\ChannelProfile.cs" />
+    <Compile Include="ShapeProfiles\CircleProfile.cs" />
+    <Compile Include="ShapeProfiles\Enums\ShapeType.cs" />
+    <Compile Include="ShapeProfiles\FabricatedBoxProfile.cs" />
+    <Compile Include="ShapeProfiles\FabricatedISectionProfile.cs" />
+    <Compile Include="ShapeProfiles\FreeFormProfile.cs" />
+    <Compile Include="ShapeProfiles\GeneralisedFabricatedBoxProfile.cs" />
+    <Compile Include="ShapeProfiles\GeneralisedTSectionProfile.cs" />
+    <Compile Include="ShapeProfiles\IProfile.cs" />
+    <Compile Include="ShapeProfiles\ISectionProfile.cs" />
+    <Compile Include="ShapeProfiles\KiteProfile.cs" />
+    <Compile Include="ShapeProfiles\RectangleProfile.cs" />
+    <Compile Include="ShapeProfiles\TSectionProfile.cs" />
+    <Compile Include="ShapeProfiles\TubeProfile.cs" />
+    <Compile Include="ShapeProfiles\ZSectionProfile.cs" />
     <Compile Include="Solid\BoundaryRepresentation.cs" />
     <Compile Include="Solid\Cone.cs" />
     <Compile Include="Solid\Cuboid.cs" />
