+++ Geometry_oM/ShapeProfiles/CircleProfile.cs
@@ -26,28 +26,16 @@ using System.Collections.ObjectModel;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace BH.oM.Structure.Properties.Section.ShapeProfiles
+namespace BH.oM.Geometry.ShapeProfiles
 {
-    public class ISectionProfile : BHoMObject, IProfile, IImmutable
+    public class CircleProfile : BHoMObject, IProfile, IImmutable
     {
-
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
+        public ShapeType Shape { get; } = ShapeType.Circle;
 
-        public ShapeType Shape { get; } = ShapeType.ISection;
-
-        public double Height { get; }
-
-        public double Width { get; }
-
-        public double WebThickness { get; }
-
-        public double FlangeThickness { get; }
-
-        public double RootRadius { get; }
-
-        public double ToeRadius { get; }
+        public double Diameter { get; }
 
         public ReadOnlyCollection<ICurve> Edges { get; }
 
@@ -55,16 +43,12 @@ namespace BH.oM.Structure.Properties.Section.ShapeProfiles
         /**** Constructors                              ****/
         /***************************************************/
 
-        public ISectionProfile(double height, double width, double webthickness, double flangeThickness, double rootRadius, double toeRadius, IEnumerable<ICurve> edges)
+        public CircleProfile(double diameter, IEnumerable<ICurve> edges)
         {
-            Height = height;
-            Width = width;
-            WebThickness = webthickness;
-            FlangeThickness = flangeThickness;
-            RootRadius = rootRadius;
-            ToeRadius = toeRadius;
+            Diameter = diameter;
             Edges = new ReadOnlyCollection<ICurve>(edges.ToList());
         }
+
 
         /***************************************************/
     }
