+++ Geometry_oM/ShapeProfiles/FreeFormProfile.cs
@@ -23,13 +23,30 @@
 using BH.oM.Base;
 using BH.oM.Geometry;
 using System.Collections.ObjectModel;
+using System.Collections.Generic;
+using System.Linq;
 
-namespace BH.oM.Structure.Properties.Section.ShapeProfiles
+namespace BH.oM.Geometry.ShapeProfiles
 {
-    public interface IProfile  : IBHoMObject
+    public class FreeFormProfile : BHoMObject, IProfile, IImmutable
     {
-        ShapeType Shape { get; }
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
+        public ShapeType Shape { get; } = ShapeType.FreeForm;
 
-        ReadOnlyCollection<ICurve> Edges { get; }
+        public ReadOnlyCollection<ICurve> Edges { get; }
+
+        /***************************************************/
+        /**** Constructors                              ****/
+        /***************************************************/
+
+        public FreeFormProfile(IEnumerable<ICurve> edges)
+        {
+            Edges = new ReadOnlyCollection<ICurve>(edges.ToList());
+        }
+
+
+        /***************************************************/
     }
 }
