+++ Geometry_oM/ShapeProfiles/IProfile.cs
@@ -23,33 +23,13 @@
 using BH.oM.Base;
 using BH.oM.Geometry;
 using System.Collections.ObjectModel;
-using System.Collections.Generic;
-using System.Linq;
 
-namespace BH.oM.Structure.Properties.Section.ShapeProfiles
+namespace BH.oM.Geometry.ShapeProfiles
 {
-    public class CircleProfile : BHoMObject, IProfile, IImmutable
+    public interface IProfile  : IBHoMObject
     {
-        /***************************************************/
-        /**** Properties                                ****/
-        /***************************************************/
-        public ShapeType Shape { get; } = ShapeType.Circle;
+        ShapeType Shape { get; }
 
-        public double Diameter { get; }
-
-        public ReadOnlyCollection<ICurve> Edges { get; }
-
-        /***************************************************/
-        /**** Constructors                              ****/
-        /***************************************************/
-
-        public CircleProfile(double diameter, IEnumerable<ICurve> edges)
-        {
-            Diameter = diameter;
-            Edges = new ReadOnlyCollection<ICurve>(edges.ToList());
-        }
-
-
-        /***************************************************/
+        ReadOnlyCollection<ICurve> Edges { get; }
     }
 }
