+++ Structure_oM/Structure_oM.csproj
@@ -120,22 +120,6 @@
     <Compile Include="Properties\Section\ConcreteSection.cs" />
     <Compile Include="Properties\Section\ExplicitSection.cs" />
     <Compile Include="Properties\Section\ISectionProperty.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\AngleProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\BoxProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\ChannelProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\CircleProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\FabricatedBoxProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\FabricatedISectionProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\FreeFormProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\GeneralisedFabricatedBoxProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\GeneralisedTSectionProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\IProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\ISectionProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\KiteProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\RectangleProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\TSectionProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\TubeProfile.cs" />
-    <Compile Include="Properties\Section\ShapeProfiles\ZSectionProfile.cs" />
     <Compile Include="Properties\Section\Reinforcement\LayerReinforcement.cs" />
     <Compile Include="Properties\Section\Reinforcement\PerimeterReinforcement.cs" />
     <Compile Include="Properties\Section\Reinforcement\Reinforcement.cs" />
